refactor: update OntologyTerm imports to use the new ontology location
