+++ src/js/types/ontology.ts
@@ -1,0 +1,4 @@
+export interface OntologyTerm {
+  id: string;
+  label: string;
+}
